+++ tests/core/console/CommandTest.php
@@ -28,5 +28,25 @@ class CommandTest extends LuyaConsoleTestCase
     public function testHelper()
     {
         $cmd = new SubCommand('myid', Yii::$app->getModule('unitmodule'));
+        $help = $cmd->getHelp();
+        $this->assertSame('', $help);
+    }
+
+    public function testPrintableDebugMessage()
+    {
+        $cmd = new SubCommand('myid', Yii::$app->getModule('unitmodule'));
+        $this->assertSame('foobar', $cmd->printableMessage('foobar'));
+        $this->assertSameTrimmed('array (
+  \'foo\' => \'bar\',           
+)', $cmd->printableMessage(['foo' => 'bar']));
+
+        $cmd->verbose = 1;
+        $cmd->verbosePrint('verbose!');
+
+        $this->app->mute = 0;
+
+        $this->assertSame(1, $cmd->outputError('test'));
+        $this->assertSame(0, $cmd->outputSuccess(['foo' => 'bar']));
+        $this->assertSame(0, $cmd->outputInfo('info'));
     }
 }